requirements.txt
certifi==2020.12.5
chardet==3.0.4
Django>=2.1.9
idna==2.8
lxml==4.9.1
pyang==2.5.3
pytz==2018.9
requests==2.26.0
urllib3==1.26.5
xym==0.5
Jinja2==3.0.3
gunicorn==20.0.4
gevent==20.9.0
piwikapi==0.3
django-ipware==4.0.2
